simple chat: sync webview config and auth status on update
