+++ vscode/src/chat/chat-view/ChatPanelsManager.ts
@@ -10,11 +10,13 @@ import { EmbeddingsSearch } from '@sourcegraph/cody-shared/src/embeddings'
 import { featureFlagProvider } from '@sourcegraph/cody-shared/src/experimentation/FeatureFlagProvider'
 
 import { View } from '../../../webviews/NavBar'
+import { getFullConfig } from '../../configuration'
 import { LocalEmbeddingsController } from '../../local-context/local-embeddings'
 import { logDebug } from '../../log'
+import { getProcessInfo } from '../../services/LocalAppDetector'
 import { createCodyChatTreeItems } from '../../services/treeViewItems'
 import { TreeViewProvider } from '../../services/TreeViewProvider'
-import { AuthStatus } from '../protocol'
+import { AuthStatus, ConfigurationSubsetForWebview, LocalEnv } from '../protocol'
 
 import { CodyChatPanelViewType } from './ChatManager'
 import { ChatPanelProvider, ChatPanelProviderOptions, ChatViewProviderWebview } from './ChatPanelProvider'
@@ -42,6 +44,7 @@ export interface IChatPanelProvider extends vscode.Disposable {
     restoreSession(chatIDj: string): Promise<void>
     setConfiguration?: (config: Config) => void
     revive: (panel: vscode.WebviewPanel, chatID: string) => Promise<void>
+    syncWebviewConfig: (authStatus: AuthStatus, configForWebview: ConfigurationSubsetForWebview & LocalEnv) => void
 }
 
 export class ChatPanelsManager implements vscode.Disposable {
@@ -92,6 +95,18 @@ export class ChatPanelsManager implements vscode.Disposable {
 
         await vscode.commands.executeCommand('setContext', CodyChatPanelViewType, authStatus.isLoggedIn)
         await this.updateTreeViewHistory()
+
+        // Update all the webview panels with the latest config, chat models, and authStatus
+        Array.from(this.panelProvidersMap.values()).map(async provider => {
+            const config = await getFullConfig()
+            const localProcess = getProcessInfo()
+            const configForWebview: ConfigurationSubsetForWebview & LocalEnv = {
+                ...localProcess,
+                debugEnable: config.debugEnable,
+                serverEndpoint: config.serverEndpoint,
+            }
+            provider.syncWebviewConfig(authStatus, configForWebview)
+        })
     }
 
     public async getChatPanel(): Promise<IChatPanelProvider> {
